+++ src/main/frontend/src/novel/novelSearch/NaverSearchResult.jsx
@@ -1,11 +1,13 @@
 import React, {useEffect, useState} from 'react';
-import {Link} from "react-router-dom";
+import {Link, useNavigate} from "react-router-dom";
 import axios from "axios";
+import {fetchData} from "../../common/NovelDetailFetch2";
 
 function NaverSearchResult(props) {
   // const [searchWord, setSearchWord] = useState(props.keyword);
   
   const [novelSearchList, setNovelSearchList]= useState([]);
+  const navi = useNavigate();
   
   useEffect(() => {
     axios.get('/searchNaver', {
@@ -52,40 +54,44 @@ function NaverSearchResult(props) {
         novelSearchList.length != 0
         ? novelSearchList.map((item, index) => {
           return (
-            <div className={'row my-4 border-top border-bottom py-2'} key={index}>
-              <div className={'col-sm-2'}>
-                <Link to={`/novelDetail?platformId=${item.platformId}&title=${item.title}&ebookCheck=${item.ebookCheck}`} >
-                  <img src={item.thumbnail} alt="" className={'w-100 h-100'} />
-                </Link>
-              </div>
-              <div className={'col-sm-10'}>
-                <Link to={`/novelDetail?platformId=${item.platformId}&title=${item.title}&ebookCheck=${item.ebookCheck}`} className={'text-decoration-none text-black fs-5 fw-bold'}>{item.title} <span className={'text-danger'}>{item.ageGrade == "Y" ? "[성인]" : null}</span>
-                </Link><br/>
-                <p className={'search-info'}>작가 : {item.author}</p>
-
-                {
-                  item.price != null
-                    ? <p className={'search-info search-price fw-bold'}>가격 : {item.price}</p>
-                    : <p className={'search-info search-price text-muted'}>{null}</p>
-                }
-                <div className={'d-flex'}>
-                  <p className={'search-info'}>
-                    평점 <span className={'fw-bold search-score'}>{item.starRate}</span>&nbsp;|
-                  </p>
-                  <p>
-                    &nbsp;<span className={'fw-bold'}>{item.completeYn}
-                    {
-                      item.title.indexOf('[단행본]') == -1
-                      ? <span>&nbsp;(총{item.count}화)</span>
-                      : <span>&nbsp;(총{item.count}권)</span>
-                    }
-                  </span>
-                  </p>
+            <Link onClick={e => {
+              fetchData(item.platformId, item.title, item.ebookCheck);
+              let fetchDataResult = fetchData(e, item.platformId, item.title, item.ebookCheck);
+              navi('/novelDetail', {state: {novelDetail: fetchDataResult}});
+            }
+            } key={index}>
+              <div className={'row my-4 border-top border-bottom py-2'}>
+                <div className={'col-sm-2'}>
+                  <img src={item.thumbnail} alt={item.title} className={'w-100 h-100'} />
                 </div>
-                <p className={'mt-3'}>{item.description.substring(0, 300)}</p>
-                
+                <div className={'col-sm-10'}>
+                  <p className={'text-decoration-none text-black fs-5 fw-bold'}>
+                    {item.title} <span className={'text-danger'}>{item.ageGrade == "Y" ? "[성인]" : null}</span>
+                  </p><br/>
+                  <p className={'search-info'}>작가 : {item.author}</p>
+                  {
+                    item.price != null
+                      ? <p className={'search-info search-price fw-bold'}>가격 : {item.price}</p>
+                      : <p className={'search-info search-price text-muted'}>{null}</p>
+                  }
+                  <div className={'d-flex'}>
+                    <p className={'search-info'}>
+                      평점 <span className={'fw-bold search-score'}>{item.starRate}</span>&nbsp;|
+                    </p>
+                    <p>
+                      &nbsp;<span className={'fw-bold'}>{item.completeYn}
+                      {
+                        item.title.indexOf('[단행본]') == -1
+                          ? <span>&nbsp;(총{item.count}화)</span>
+                          : <span>&nbsp;(총{item.count}권)</span>
+                      }
+                  </span>
+                    </p>
+                  </div>
+                  <p className={'mt-3'}>{item.description.substring(0, 300)}</p>
+                </div>
               </div>
-            </div>
+            </Link>
           )
         })
         : <div className={'d-flex justify-content-center'}>
